requirements.txt
Flask==1.1.2
flask_jwt_extended==4.4.4
pymongo==4.3.3
python-dotenv==0.21.1
